Add files via upload
Hey look, I made a game! Okay, I followed the directions on a code school course. I need to add some functionality to ask the users if they even want to play the game, as well as actually make it work when I ask if they want to try again.